Build automatization.
git-svn-id: https://cvet.by:8443/svn/FOnlineSource/Main@1189 c2c6a35e-ad2b-4e1f-bb64-97b989be360b

Former-commit-id: 78ba1329564e47663ebf9d9ae976f849433ff61d

+++ Source/ClientInterface.cpp
@@ -512,7 +512,7 @@ void FOClient::FormatTags( char(&text)[ MAX_FOTEXT ], CritterCl* player, Critter
                     rnd.pop_back();
                 if( !rnd.empty() )
                 {
-                    char* rnd_str = rnd[ Random( 0, rnd.size() - 1 ) ];
+                    char* rnd_str = rnd[ Random( 0, (int) rnd.size() - 1 ) ];
                     str_ = rnd_str + 1;
                     Str::GoTo( str_, '|' );
                     Str::CopyCount( tag, rnd_str + 1, (uint) ( str_ - rnd_str ) - 1 );
